+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+- 0.3.0 (2025/6/5)
+
+  - Change "Download Markdown" feature to "Copy Mindmap Content" for a mindmap.
+  - Add multilingual support for all the buttons.
+
 - 0.2.0 (2025/6/5)
 
   - Add a new button to "Download Markdown" for a mindmap.